dicts moved to parser_id
